+++ tests/testthat/test-names.R
@@ -27,3 +27,13 @@ test_that("names_sort() works", {
   expect_error(sort_names(list(a = 1)))
   expect_error(sort_names(NA))
 })
+
+test_that("%names% works", {
+  x <- 1:4
+  nm <- letters[1:4]
+
+  expect_identical(
+    x %names% nm,
+    set_names0(x, nm)
+  )
+})